iplug-dsc-mapclient: -intial import
This Iplug, at this point, takes the configuration file of the webmapclient and fetches the capabilities urls. 
These urls are requested, the resulting wms-documents are mapped into a lucene index, with fields named according to their usual database fields. In a second step these field are mapped into an idf document, which also gets stored in the lucene index. This is done via the usual mapping technique(js-mapping). It has a configuration menu and the application.properties file known from the other dsc.
On a request from the portal it will return a detailed view via the idf stored in the index.